backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
Pillow>=10.0
ffmpeg-python>=0.2.0
